feat: Add openapi and root routes
